refactor: move exposed __fixtures__ into a mocks folder
this will help us separate mocks we want to expose to the world VS internal types for testing.
