+++ packages/profile-sync-controller/auth/mocks/package.json
@@ -4,6 +4,6 @@
   "description": "",
   "license": "MIT",
   "sideEffects": false,
-  "main": "../../dist/controllers/authentication/__fixtures__/index.cjs",
-  "types": "../../dist/controllers/authentication/__fixtures__/index.d.cts"
+  "main": "../../dist/controllers/authentication/mocks/index.cjs",
+  "types": "../../dist/controllers/authentication/mocks/index.d.cts"
 }
