+++ packages/profile-sync-controller/jest.config.js
@@ -27,6 +27,7 @@ module.exports = merge(baseConfig, {
   coveragePathIgnorePatterns: [
     ...baseConfig.coveragePathIgnorePatterns,
     '/__fixtures__/',
+    '/mocks/',
     'index.ts',
   ],
