+++ packages/profile-sync-controller/package.json
@@ -49,12 +49,12 @@
     },
     "./auth/mocks": {
       "import": {
-        "types": "./dist/controllers/authentication/__fixtures__/index.d.mts",
-        "default": "./dist/controllers/authentication/__fixtures__/index.mjs"
+        "types": "./dist/controllers/authentication/mocks/index.d.mts",
+        "default": "./dist/controllers/authentication/mocks/index.mjs"
       },
       "require": {
-        "types": "./dist/controllers/authentication/__fixtures__/index.d.cts",
-        "default": "./dist/controllers/authentication/__fixtures__/index.cjs"
+        "types": "./dist/controllers/authentication/mocks/index.d.cts",
+        "default": "./dist/controllers/authentication/mocks/index.cjs"
       }
     },
     "./user-storage": {
@@ -69,12 +69,12 @@
     },
     "./user-storage/mocks": {
       "import": {
-        "types": "./dist/controllers/user-storage/__fixtures__/index.d.mts",
-        "default": "./dist/controllers/user-storage/__fixtures__/index.mjs"
+        "types": "./dist/controllers/user-storage/mocks/index.d.mts",
+        "default": "./dist/controllers/user-storage/mocks/index.mjs"
       },
       "require": {
-        "types": "./dist/controllers/user-storage/__fixtures__/index.d.cts",
-        "default": "./dist/controllers/user-storage/__fixtures__/index.cjs"
+        "types": "./dist/controllers/user-storage/mocks/index.d.cts",
+        "default": "./dist/controllers/user-storage/mocks/index.cjs"
       }
     },
     "./package.json": "./package.json"
