+++ packages/profile-sync-controller/src/controllers/authentication/index.ts
@@ -4,4 +4,4 @@ const AuthenticationController = Controller;
 export { Controller };
 export default AuthenticationController;
 export * from './AuthenticationController';
-export * as Mocks from './__fixtures__';
+export * as Mocks from './mocks';
