+++ packages/profile-sync-controller/src/controllers/authentication/mocks/index.ts
