+++ packages/profile-sync-controller/src/controllers/authentication/mocks/mockResponses.ts
