+++ packages/profile-sync-controller/src/controllers/user-storage/__fixtures__/mockServices.ts
@@ -1,6 +1,11 @@
 import nock from 'nock';
 
 import {
+  USER_STORAGE_FEATURE_NAMES,
+  type UserStoragePathWithFeatureAndKey,
+  type UserStoragePathWithFeatureOnly,
+} from '../../../shared/storage-schema';
+import {
   getMockUserStorageGetResponse,
   getMockUserStoragePutResponse,
   getMockUserStorageAllFeatureEntriesResponse,
@@ -8,12 +13,7 @@ import {
   getMockUserStorageBatchDeleteResponse,
   deleteMockUserStorageAllFeatureEntriesResponse,
   deleteMockUserStorageResponse,
-} from './mockResponses';
-import {
-  USER_STORAGE_FEATURE_NAMES,
-  type UserStoragePathWithFeatureAndKey,
-  type UserStoragePathWithFeatureOnly,
-} from '../../../shared/storage-schema';
+} from '../mocks/mockResponses';
 
 type MockReply = {
   status: nock.StatusCode;
