+++ packages/profile-sync-controller/src/controllers/user-storage/__fixtures__/test-utils.ts
@@ -1,7 +1,7 @@
 import type nock from 'nock';
 
-import { MOCK_STORAGE_KEY } from './mockStorage';
 import encryption from '../../../shared/encryption/encryption';
+import { MOCK_STORAGE_KEY } from '../mocks/mockStorage';
 import type {
   GetUserStorageAllFeatureEntriesResponse,
   GetUserStorageResponse,
