+++ packages/profile-sync-controller/src/controllers/user-storage/index.ts
@@ -4,6 +4,6 @@ const UserStorageController = Controller;
 export { Controller };
 export default UserStorageController;
 export * from './UserStorageController';
-export * as Mocks from './__fixtures__';
+export * as Mocks from './mocks';
 export * from '../../shared/encryption';
 export * from '../../shared/storage-schema';
