+++ packages/profile-sync-controller/tsconfig.build.json
@@ -12,5 +12,11 @@
     { "path": "../accounts-controller/tsconfig.build.json" },
     { "path": "../network-controller/tsconfig.build.json" }
   ],
-  "include": ["../../types", "./src"]
+  "include": ["../../types", "./src"],
+  "exclude": [
+    "./jest.config.packages.ts",
+    "**/*.test.ts",
+    "**/jest.config.ts",
+    "**/__fixtures__/"
+  ]
 }
