+++ packages/profile-sync-controller/user-storage/mocks/package.json
@@ -4,6 +4,6 @@
   "description": "",
   "license": "MIT",
   "sideEffects": false,
-  "main": "../../dist/controllers/user-storage/__fixtures__/index.cjs",
-  "types": "../../dist/types/controllers/user-storage/__fixtures__/index.d.cts"
+  "main": "../../dist/controllers/user-storage/mocks/index.cjs",
+  "types": "../../dist/types/controllers/user-storage/mocks/index.d.cts"
 }
